### 2021-03-19 #35 @hellosakshi
Submission: easysklearn (Python)
Submitting Author: Name (@github_handle)

- Lara habashy (@larahabashy)
- Ifeanyi Anene (@IfyAnene7)
- Sakshi Jain (@hellosakshi)
- Zhenrui (Eric) Yu (@yzr1996)

Package Name: easysklearn

One-Line Description of Package: A python package to explore data, to select features, to impute missing data and to select the model

Repository Link:  https://github.com/UBC-MDS/524_easysklearn

Version submitted:
Version accepted: TBD

---

`Easysklearn` is a python package designed to perform exploratory data analysis, to help with missing data imputation and to give baseline models. Also, it assists in feature selection which is a common problem when undertaking a data science or machine learning analysis.
	- [x]  Exploratory data analysis
	- [x] Feature selection
	- [x] Model selection
	- [x] Missing data imputation

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):
It explores data and impute missing values. Also, it helps in feature selection and helps in selecting the model.

The target audience of this package includes anyone who has the requirements to clean data and build machine learning model. For instance, students interested in machine learning might be the target audience. Also, data scientists, data engineers, statisticians can be possible target users.

There are some Python packages to explore data (EDA) such as "Pandas Profiling". However, our Python package provides a different working flow and functionality. There are four main components of our Python package and each of the functions has its innovation points. For example, our eda function gives users a sense of the whole distribution summary of the raw data. In addition to EDA, our model helps in filling the missing values, in feature selection and in selecting the best model.

NA

- [ ] has an [OSI approved license][OsiApprovedLicense].

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
