### 2021-03-19 #35 @NickleDave
Never mind I think I figured out! :)

https://github.com/UBC-MDS/software-review-2021

Happy to provide material so you all can get practice submitting for review
